use icon from icon theme
This commit makes the app try to use the icons from the current icon
theme, and falls back to the bundled ones if none is available.

Icon recoloring is not supported, so themes relying on it (e.g. breeze
dark) will look broken (dark icons on dark background). Note that the
current implementation always use dark icons even when using a dark
color scheme.